Revert "move chunker to borg.algorithms"
This reverts commit 956b50b29cb6d3aec6b7a02d46325e8ab50bc149.

# Conflicts:
#	setup.py
#	src/borg/archive.py
#	src/borg/helpers.py
