+++ .gitignore
@@ -8,7 +8,7 @@ src/borg/compress.c
 src/borg/crypto/low_level.c
 src/borg/hashindex.c
 src/borg/item.c
-src/borg/algorithms/chunker.c
+src/borg/chunker.c
 src/borg/algorithms/checksums.c
 src/borg/platform/darwin.c
 src/borg/platform/freebsd.c
